requirements.txt
spotipy==2.9.0
PyYAML==5.3